download button's functionality fixed
