+++ src/features/bbps/components/transaction-search.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import { IconDeviceMobile, IconCalendar } from '@tabler/icons-react'
+import { IconDeviceMobile, IconCalendar, IconDownload } from '@tabler/icons-react'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import {
@@ -9,6 +9,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components/ui/select'
+import axios from 'axios'
 
 interface TransactionSearchProps {
   onSearch: (params: {
@@ -39,6 +40,12 @@ export function TransactionSearch({
     paymentStatus: '',
   })
 
+  const BACKEND_BASE_URL = import.meta.env.VITE_BACKEND_BASE_URL;
+  const getToken = () => {
+    const match = document.cookie.match(/(?:^|; )auth_token=([^;]*)/)
+    return match ? decodeURIComponent(match[1]) : ''
+  }
+
   const handleChange = (
     e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>
   ) => {
@@ -61,6 +68,55 @@ export function TransactionSearch({
       paymentStatus: '',
     })
     onReset()
+  }
+
+  const handleDownload = async () => {
+    const token = getToken();
+    const params = {
+      ...(fields.id && { transaction_id: fields.id }),
+      ...(fields.bbpsReferenceCode && { bbps_ref_no: fields.bbpsReferenceCode }),
+      ...(fields.customerId && { customer_id: fields.customerId }),
+      ...(fields.paymentStatus && { status: fields.paymentStatus }),
+      ...(fields.mobileNumber && { mobile: fields.mobileNumber }),
+      ...(fields.category && { category: fields.category }),
+      ...(fields.start_date && { start_date: fields.start_date }),
+      ...(fields.end_date && { end_date: fields.end_date }),
+      isDownload: true,
+    };
+    try {
+      const response = await axios.get(
+        `${BACKEND_BASE_URL}/v1/bbps/getAllTransactions`,
+        {
+          params,
+          headers: { Authorization: `Bearer ${token}` },
+          responseType: 'blob',
+          withCredentials: true,
+        }
+      );
+      let filename = 'bbps-transactions.xlsx';
+      const disposition = response.headers['content-disposition'];
+      if (disposition) {
+        const match = disposition.match(/filename[^;=\n]*=((['"]).*?\2|[^;\n]*)/);
+        if (match && match[1]) {
+          filename = match[1].replace(/['"]/g, '');
+        }
+      }
+      const blob = new Blob([response.data]);
+      if (blob.size < 2000) {
+        alert('No transactions found for the selected range.');
+        return;
+      }
+      const url = window.URL.createObjectURL(blob);
+      const link = document.createElement('a');
+      link.href = url;
+      link.setAttribute('download', filename);
+      document.body.appendChild(link);
+      link.click();
+      link.parentNode?.removeChild(link);
+      window.URL.revokeObjectURL(url);
+    } catch {
+      alert('Failed to download Excel.');
+    }
   }
 
   return (
@@ -110,7 +166,7 @@ export function TransactionSearch({
               name='mobileNumber'
               value={fields.mobileNumber}
               onChange={handleChange}
-              placeholder='Enter Mobile Number'
+              placeholder='Enter Mobile Number' 
               className='pl-8'
             />
           </div>
@@ -200,6 +256,16 @@ export function TransactionSearch({
           onClick={handleReset}
         >
           Reset
+        </Button>
+        <Button
+          type='button'
+          variant='outline'
+          className='h-9'
+          onClick={handleDownload}
+          disabled={Object.values(fields).every((value) => value === '')}
+        >
+          <IconDownload className='mr-1' />
+          Download Report
         </Button>
       </div>
     </form>
